# wallet-ui/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "数字钱包demo服务端",
        "description": "描述...",
        "contact": {
            "name": "zjh",
            "email": "zjh181240@aliyun.com"
        },
        "version": "版本号: 1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/user/{userId}/roles/{roleId}": {
            "get": {
                "operationId": "getLogin",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/javabeat/{regexp1}": {
            "get": {
                "operationId": "getRegExp",
                "parameters": [
                    {
                        "name": "regexp1",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/save_user": {
            "get": {
                "operationId": "saveUser",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "saveUser_3",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "saveUser_2",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "saveUser_5",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "options": {
                "operationId": "saveUser_6",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "head": {
                "operationId": "saveUser_1",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "saveUser_4",
                "parameters": [
                    {
                        "name": "u",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_2",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "options": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "head": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "hello接口",
                "description": "hello接口",
                "operationId": "hello_4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名称",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "unknown user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "欢迎字符串",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "description": ""
            }
        }
    }
}